split up files tests
